feat: add reusable ScrollContainer and enhance PageHeader
- Created `ScrollContainer` component with auto-height and custom scrollbar using Tailwind
- Enhanced `PageHeader` with `subtitle` and `extraContent` props to support secondary info (e.g., Attribute ID, displayType badge)
- Updated Attribute page to use new PageHeader props and ScrollContainer
