+++ app/app.js
@@ -23,7 +23,7 @@ import { SubSubCategoryPage } from './containers/Products/sub sub category';
 import SubSubCategoryFormPage from './containers/Products/sub sub category/SubSubCategoryFormPage';
 import SubSubCategoryView from './containers/Products/sub sub category/SubSubCategoryView';
 import Brand from './containers/Products/Brand/BrandPage';
-import Attributes from './containers/Products/Attributes/AttributePage';
+import Attributes from './containers/Products/attribute/AttributePage';
 import ProductInHouse from './containers/Products/Product/ProductInHouse';
 import Vendors from './containers/Products/Ventors/Vendors';
 import BulkImpor from './containers/Products/Bulk/BulkImpor';
@@ -57,8 +57,8 @@ import Payment_method from './containers/Business_Setting/Payment_method';
 import Seo_setting from './containers/Business_Setting/Seo_setting';
 import './styles/toast.css';
 import 'react-toastify/dist/ReactToastify.css';
-import AttributeformPage from './containers/Products/Attributes/AttributeformPage';
-import AttributeView from './containers/Products/Attributes/AttributeView';
+import AttributeformPage from './containers/Products/attribute/AttributeformPage';
+import AttributeView from './containers/Products/attribute/AttributeView';
 import BrandFormPage from './containers/Products/Brand/BrandFormPage';
 import BrandView from './containers/Products/Brand/BrandView';
 // import './styles/toast.css';
