+++ app/containers/Products/sub sub category/SubSubCategoryPage.jsx
@@ -1,9 +1,7 @@
-import { Button, Container, Breadcrumb } from '@/components';
+import { Button, Container, Breadcrumb, PageHeader } from '@/components';
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import SubSubCategoryList from './SubSubCategoryList';
-
-
 
 const SubSubCategoryPage = () => {
   const [refreshTrigger, setRefreshTrigger] = useState(0);
@@ -15,30 +13,15 @@ const SubSubCategoryPage = () => {
 
   return (
     <Container>
-      <div className="mb-8">
-        <div className="flex items-center justify-between">
-          <div>
-            <h1 className="text-3xl font-bold text-gray-900">
-              Sub Sub Category
-            </h1>
-          </div>
-          <Button
-            onClick={handleAddSubSubCategory}
-            variant="primary"
-            className="flex items-center space-x-2"
-          >
-            <span>+</span>
-            <span>Add Sub Sub Category</span>
-          </Button>
-        </div>
-        <Breadcrumb
-          items={[
-            { label: 'Dashboard', href: '/dashboard' },
-            { label: 'Sub Sub Category' },
-          ]}
-        />
-      </div>
-
+      <PageHeader
+        title="Sub Sub Category"
+        breadcrumbItems={[
+          { label: 'Dashboard', href: '/dashboard' },
+          { label: 'Sub Sub Category' },
+        ]}
+        onAddClick={handleAddSubSubCategory}
+        addButtonLabel="Sub Sub Category"
+      />
       <SubSubCategoryList refreshTrigger={refreshTrigger} />
     </Container>
   );
