+++ app/containers/Products/sub sub category/SubSubCategoryView.jsx
@@ -1,11 +1,16 @@
-import { Button, LoadingData, Container, Breadcrumb, CustomIcon } from '@/components';
+import {
+  Button,
+  LoadingData,
+  Container,
+  Breadcrumb,
+  CustomIcon,
+  PageHeaderWithActions,
+} from '@/components';
 import React, { useState, useEffect } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
 
-
 import categoryService from '@/api/service/categoryService';
 import { toast } from 'react-toastify';
-
 
 const SubSubCategoryView = () => {
   const { id } = useParams();
@@ -148,41 +153,28 @@ const SubSubCategoryView = () => {
 
   return (
     <Container>
-      <div className="mb-8">
-        <div className="flex items-center justify-between">
-          <div>
-            <h1 className="text-3xl font-bold text-gray-900">
-              Sub-Sub-Category Details
-            </h1>
-          </div>
-          <div className="flex space-x-3">
-            <Button
-              variant="outline"
-              onClick={handleEdit}
-              className="flex items-center space-x-2"
-            >
-              <CustomIcon type="edit" size={4} />
-              <span>Edit Sub-Sub-Category</span>
-            </Button>
-            <Button
-              variant="secondary"
-              onClick={handleBack}
-              className="flex items-center space-x-2"
-            >
-              <CustomIcon type="arrow-left" size={4} />
-              <span>Back to Sub-Sub-Categories</span>
-            </Button>
-          </div>
-        </div>
-        <Breadcrumb
-          items={[
-            { label: 'Dashboard', href: '/dashboard' },
-            { label: 'Sub-Sub-Categories', href: '/products/subsubcategories' },
-            { label: subSubCategory.name },
-          ]}
-        />
-      </div>
-
+      <PageHeaderWithActions
+        title="Sub Sub Category Details"
+        breadcrumbItems={[
+          { label: 'Dashboard', href: '/dashboard' },
+          { label: 'Sub-Sub-Categories', href: '/products/subsubcategories' },
+          { label: subSubCategory.name },
+        ]}
+        actions={[
+          {
+            label: 'Edit Sub Sub Category',
+            onClick: handleEdit,
+            variant: 'outline',
+            icon: <CustomIcon type="edit" size={4} />,
+          },
+          {
+            label: 'Back',
+            onClick: handleBack,
+            variant: 'secondary',
+            icon: <CustomIcon type="arrow-left" size={4} />,
+          },
+        ]}
+      />
       <div className="bg-white rounded-lg shadow overflow-hidden">
         <div className="px-6 py-4 border-b border-gray-200">
           <h2 className="text-xl font-semibold text-gray-900">
